Always use normalized arguments from the shared validator
Only very few normalizations are currently in place.
* dir is lower-cased.
* group gets a default.
* details is dropped in invalid contexts.

The first two mean that we can remove the duplicate normalization
code.

The last means that Parsoid will stop rendering details in contexts
where it's not allowed. All affected test cases are about this. The
output is much more in line with the legacy parser now.

Note that round-trips are not affected.

Bug: T393022
Change-Id: I042e15515e80407b12b3160736565f6ec6176c97