Static ref preview when editing details
Drop the subreference warning.  Show a preview of the main ref and a
static placeholder where the details will appear.

This is an incremental step and we'll try to implement a live preview
of actual contents in future work.

Short explanations of some strangeness in this patch:
* The edit panel is split into two layouts to allow a full-width
divider between the preview and other content.
* <li> numbering is done manually to get better control over
the left margin in both desktop and mobile views.

Bug: T394480
Change-Id: Ie686a1e57451be881e7ed3025c7af1f1f49442be