Add edit button for subrefs in context item
Added a second edit button for subrefs for the context item which
appears when clicking on a footnote marker in VE edit mode.

Wired to two behaviors: the top button edits the main ref, the bottom
button edits the subref.

Parameters and i18n inspired from the ve.ui.LinearContextItem class.

Bug: T394481
Change-Id: I0b6030616665d3a7f9f6780026fd170e2c7f634c