Revert and add test for "in indicator context" logic
The patch Idaf19900b8e65956c74f165cd8b8a80ea1871b35 changed the
logic of being in an "indicator context" from checking only the
last level to checking all levels. This was possible because it
was previously not covered by any tests, neither unit nor parser
tests.

This patch reverts this back to checking only the last level, as
well as adds the missing test. This fixes one of the two issues
described in T393913.

Bug: T393913
Change-Id: Idd3f394fef644c47575ccee449fe3d6c225ae660