Show reuse chip for main part of subref
The total usage count for a subref should come from the main ref, not
the (unreusable) subref.

This got lost in some of the sub-tickets we created but can be seen
in T392687, especially in https://phabricator.wikimedia.org/F59961869.

TODO: Note that the chip appears in the wrong place, so preview
blocks need to be rearranged in separate work.

Bug: T394479
Change-Id: I84a5d4f07f496a501042c3448588096aa5c645ef