Rearrange reuse chip to appear between main and details
Switch how the main and details previews are rendered, so that the
reuse chip is attached to the main block and appears in the same
place whether or not details are present.

This got lost in some of the sub-tickets we created but can be seen
in T392687, especially in https://phabricator.wikimedia.org/F59961869.

Bug: T394479
Change-Id: I09225889506aba02777baf4689b4f3d8f975a759