Add more QUnit tests for ve.dm.MWReferenceNode
These QUnit tests are more regression tests. The setup relies on
anonymous mocks and is not super flexible because of this. Such tests
can break when unrelated changes are made, and might be hard to
update. Because of this we should feel free to remove such a test
when it turns out it's not useful any more.

Additional changes:

* I turned the String.match() calls around. The .exec() does the
same as before (as proven by the tests) but is much more robust when
the string is e.g. null or undefined.

* The `!nodesWithSameKey` check was bogus. Arrays cannot be compared
like this in JavaScript. This was effectively dead code. I removed it.

* Instead I added a safe-guard to make sure the code makes sense when
the array is empty or only contains a single ref (the current one)
anyway. Note there is a similar `nodesWithSameKey.length > 1` just a
few lines below.

Bug: T392977
Change-Id: Ie74737c1c98b27cab97c655511d1f084c6c27dd4