VE: Emit fresh reflist when saving from VE
In this patch we render a complete reflist when saving from VE to
pass on to Parsoid.  This is needed to be able to pickup changes
to the main content when a reference is used where the main node
is only part of the references list as "synthetic" ref.

Making sure this content is rendered and present in the output
allows us to pick it up when Parsoid needs to generate the updated
Wikitext.

I tried to make the change as small as possible, ignoring edge
cases that might arise with e.g. templates.  See the TODOs in the
code.

Bug: T396017
Change-Id: Iaed01fc2f0719ab36b48b03a090aac36d54ee492