Remove cite_warning wrapper message in favor of cite_error
There is no point in having two (almost) identical wrapper
messages. The only difference in the code is:
* A "warning" doesn't get the tracking category.
* It gets an extra CSS class.
But this is all dictated by the inner message, not by the wrapper.
The tests demonstrate that no HTML or CSS changes.

The concept was introduced in 2014 via I2ca1b62 for a single error
message. It was never used for anything else. Since the message
is about "section previews", but section previews don't exist in
Parsoid, it is not relevant for Parsoid and only appears in the
legacy parser.

Since then we re-introduced a similar concept of errors with
different severity (warnings vs. fatals) in the Validator class.
But this is has nothing to do with the wrapper message.

In contrast to cite_error there is no relevant customizaion for
cite_warning anywhere:
https://global-search.toolforge.org/?q=.*&regex=1&namespaces=8&title=Cite+warning.*

Bug: T7984
Bug: T353695
Change-Id: I1c59e034dc206b03189360bacde9f23c2036763b