Replace old "extendsRef" attribute name everywhere
Via T373307 we decided to use "sub-reference" and "main reference".
The word "extends" is still in the code because we reused a lot of
the existing code.

This is actually the "listKey" of the "main reference", i.e. not
the name but the long, internal string key. As such the name could
be "mainRefListKey" or just "mainKey". What do you prefer?

"mainRef" shouldn't be used. We use it already for:
* Either the name of the main ref.
* Or the actual main ref model object.

I think this patch shouldn't have any visible effect. Note that this
feature is not yet deployed anywhere.

Bug: T373307
Bug: T397145
Depends-On: I4c9027558e645a4c282c078aa33659d05a7e87a1
Change-Id: I8e574ba815c553134fb5e7cebbbc47acd0d4dc33