Parsoid: Add bodyHtml of synthetic refs from `details` to mwData
I added special handling for synthetic ref bodies so that the data
is forwarded also when the reflist is transcluded.  With this done
the data can then be read in VE to show the contents of the main
refs.

In this patch I only do the change necessary in Parsoid.  There's
a piece missing in VE then to get the contents from the mwData
body.  This is currently disabled for transclusions.

( See T209493#4886274 )

Bug: T395083
Change-Id: I4b0dfb86ec4f27689759896ddb989e4332c960ec