Fix VE meta data for sub-ref with empty main content slot
The code was able to understand a sub-ref like <ref details=… />
that does not have a main body and is not a "sub-ref with main
body".

But a semantically identical <ref details=…></ref> with an empty
string was behaving very different.

We don't need a "synthetic main ref" to track this. We also don't
want to "reserialize the main content" into such a sub-ref. It
should stay empty.

Note this doesn't fully fix all issues but is already a critical
improvement. Before the content of the details="" attribute got
lost. The remaining issue after this patch is no content loss any
more but only a minor dirty diff. The empty <ref></ref> becomes
<ref />. We can improve this later.

Bug: T392195
Change-Id: I426e2cd540a975b620718e9b1b31099effe7807c