Remove not needed implementation details from InternalList test
This test was originally written with the assumption that the
individual objects returned by .getNodeGroups() are plain objects.
This changed via I1778c91.

I had a closer look and found that this test is doing a lot of
assertions that are either exposing implementation details or are
unrelated to what the test is meant to prove. For example:
* The uniqueListKeys… properties are for an internal counter that's
  not even used in this test.
* indexOrder doesn't matter when there was nothing to order, except
  in a single, specific situation.
* firstNodes is an internal cache that repeats the first nodes from
  the keyedNodes structure. It's not used anywhere in this test.

Bug: T397395
Change-Id: I848cb77f100e2f1d7bb06ec489d7bbe1b19cc269