Fix VE rendering conflicting main instead of sub-ref
Note this is not a full fix but only a small step forward. Before
this patch the content of the sub-ref got lost. You can see this
when editing in VE. That's a very obviously broken user experience.

After this patch the bad, conflicting content gets lost. That's much
less of a problem. That conflicting content was never accessible
anyway (outside of plain wikitext editing).

We still need to fix this. Still I would argue that this is a step
forward and can already be merged as it is.

This cannot affect wikis where sub-refs are disabled.

Bug: T392814
Change-Id: I22501a58739bc9ce9fe60d21544816c2e57c8739