Ensure self-referential refs with redefinitions emit errors, not crash
The goal of this patch is to ensure robustness in the face of
self-referential refs, especially ones that attempt to redefine the
content of the ref.  The particular output, as in, which ref content
takes priority, is less important since this isn't really something
we're trying to support.

The approach in this patch is to put a lock on a named ref to prevent
any similarly named refs found in its content from setting a definition
for the name first.  We can't set the definition for a ref before
processing its content because that would change the sequence number of
ref-in-refs in general when compared to the legacy output.

In I4059e32b9bea8cdc23d2112812c3f7e167e47399, we anticipated the current
crasher.  When we're traversing looking for ref-in-ref and there's no
conflict, we assume that we aren't in an embedded context.  However, if
we encounter a ref with the same name and a definition, that would take
priority as the contentId and then we trivially have a conflict with the
current ref we're processing and need to mark its content as embedded.

An alternate implementation in I88392f0ee7a9cc4a0b7d0c9a88d2dbedc3132884
tried to detect when this occurs and go back and mark nodes as embedded
after the fact but that quickly proved unwieldy, as captured in the
tests.

Bug: T382546
Bug: T393913
Change-Id: Id4bc98544fcd8fb2ed80b4389f228b3b9cb2b48d