Parsoid: Ignore synthetic main refs from details
When the main reference is not part of it's own tag, but combined
with `details` we split these two and create a 'synthetic' list
defined ref.  This ref should never be converted to Wikitext or
lead to the creation of a references tag there.

Note that there's another patch that makes sure updated content of
such a ref is still forwarded to Parsoid.

See Iaed01fc2f0719ab36b48b03a090aac36d54ee492

Bug: T389929
Change-Id: Iaca39747bd854b4d55b81d82f32418b76d60e845