VE: Enforce referenceslist reserialization when MW changed
Here I'm applying a similar logic as done in MWReferenceNode.  At
the moment ReferencesList would not reserialize when the MW data
changed, because it does not matter for the output.  This will be
different for the upcoming patch where we want to actually send
some output.

To make sure that this is not replaced by the stored original I
added the extra check here.  See something similar done in the
MWReferenceNode converter code.

Bug: T396017
Change-Id: I6c261bff863b7a9ec8ef7e18d5dc3bf72ff23745