Re-classify "numeric <ref> name" from fatal to warning
This does have a few significant effects:

1. The <ref> tag is not destroyed any more. Only the invalid name
is dropped. The <ref> is rendered and usable for readers. The error
message still appears, as before, but at the end of the reference
in the reference list.

2. The behavior of the two parsers was completely different before.
It's now much more similar.

The existing test cases demonstrate what changes under the hood:
* In the classic parser the error messages move down into the
  reference list.
* Some global ids change and increase because there is one more
  <ref> to render and count. That's a harmless side-effect and
  expected.
* We can see that Parsoid doesn't generate invalid ids like
  "cite_note-1-2" and more. "2" is the global id but "1" is the
  invalid name that should have been dropped.

This is split from Iab8059f and changes only the name="…"
attribute, not follow="…".

Bug: T238061
Change-Id: I1389f7e58b1f7420efa26a7c0ec3cc6b577b852a