Re-classify mismatching group from fatal to warning
This is a rare error when the group="…" attribute of a
list-defined <ref> and the outer <references> doesn't match.

The main effect of this change is that it brings the two parsers
much closer together. In detail:

1. In the classic parser the confusing (and wrong) error message
about missing content disappears. Only the correct error remains.

2. Same in Parsoid. A whole bunch of misleading follow-up errors
disappear.

Parsoid renders only one reflist now, not two. That was also just
confusing and misleading.

Bug: T238061
Change-Id: I4c767d61569718efb2fe986625b08802290783d8