Move duplicate "responsive" validation into the Validator
The Validator is already doing some normalization/sanitization, see
docs/error-reporting.md. This is just another situation where we can
very easily avoid petition in the two parsers.

This is motivated by T400013 because we have seen the "responsive"
attributes in some of the edits affected by T400013. But it's most
certainly not related.

Change-Id: I246578b55ddcd8b9abeeeafd63415e6bb325546b