Unify different syntax in parser tests .txt files
The parser test parser is flexible and accepts slightly different
formats in several places, e.g. the space after !! is optional.
Let's unify the differences a little bit to make it easier to read.

The file name format *Tests.txt makes it easier to identify files
when only the file name is visible. This is e.g. relevant when we
want to assign a custom syntax highlighter in our IDEs.
Intentionally only touching a single file for the moment. We can
rename more files if it turns out this really is helpful.

Change-Id: I1b36fd6415290227a47c4f684672b1b6880ff3e2