VE: Add Converter test with list defined refs and fresh reflist
I'm adding a Convert test with list defined refs.  This one is not
taking the ReferenceList from stored originalHtml but instead emits
a fresh one according to our latest code changes.

Note that it shows the missing ids on the reflist items. That's what
lead to the bug we want to cover here.

Bug: T400052
Change-Id: Idf578a56c644d3071a2e889003636e25bc61f8df