VE: Make sure that emitted relist items have an id
Our original implementation did not take into account, that not
all nodes that belong to a key have a refListItemId.  We need that
id though so that Parsoid can safly convert back to Wikitext and
find the content.

This fixes the issues we ran into with Parsoid deleting list de-
fined refs when using our code that emits fresh reflists.

Bug: T396017
Bug: T400038
Change-Id: I7f2b3aa8d8ba5710eccef25291426f597932bc73