Parsoid: Add test to cover the minimal VE reflist output
When we worked on emiting a fresh references list in VE on save
we ran into the issue of untouched deleted list defined refs.

Here I demostrate what's needed from a minimal reflist HTML to
convert back to list defined references in Wikitext.

It seems, that the reflistId on the nodes in the reflist is
essential.  This was missing in some cases on the nodes of the
mininal reflist HTML.

Bug: T396017
Bug: T400052
Change-Id: I4f2d774bc1619e2d81da5ce296226981697aae81