Converter: Improve simple main+details and add fitting Parsoid tests
I renewed some of the basic Converter tests for details.  They were
partly broken.  I also added one way Parsoid tests for the output
of the Converter to cover what Parsoid is expected to do in these
cases.

I also used the basic setup the of the test to spin off a modified
version without having to repeat the whole internal data.  This
way we can easily test changes on a certain input without
repetition.

Bug: T400052
Bug: T389929
Change-Id: I4276deaf518b6613944f673576f58679e662a36e