[Refactor] Move VE Converter test to dedicated files
I'm untangling the test setups a bit moving the Converter cases
to it's own file.  With that I also split them in two categories.

The default cases are mostly coming what's already been there in
that repo.  The cases I call 'Integration' test cases should always
have a fitting pair of tests for Parsoid where we're using the
Converter's HTML output to make sure the Wikitext generated is
also as expected.

Bug: T400052
Change-Id: Iebe2209b3f014c8f00be11c90e38375774e5f3d7