Rename all parser test files to *Tests.txt
I started using a custom syntax highlighter for parser tests. It's
not much, but it helps. Unfortunatelly it requires some kind of
naming convention for the files to understand the difference between
normal *.txt files and which of them are parser tests. But we have
already some kind of naming convention in place, don't we? It's just
not consistent.

Change-Id: I3bf9b9d9776fd1665e276e3b55d703ec1ea4d7d0