Converter: Make sure to emit a fresh reflist when contents change
Here I'm applying a similar logic as done in MWReferenceNode.  At
the moment ReferencesList would not reserialize when the MW data
changed, because it does not matter for the output.  With details
it's a bit different.  We need to emit a fresh reflist so that
the content of synthetic list defined refs can update the Wikitext

So to make sure that we're not using the stored original HTML I
added the extra check.

Note that this re-does I6c261bff863b7a9ec8ef7e18d5dc3bf72ff23745.
That patch needed to get reverted because of vanishing list
defined refs.

We fixed that issue by making sure that all the refs in the emitted
reflist have an id. See I7f2b3aa8d8ba5710eccef25291426f597932bc73.

We also added a lot of tests that fill a gap we had when it comes
to the conversion from Converter output to Parsoid to Wikitext.
See the tickets under T400052 and the changes to tests that needed
to be done here.

Bug: T400798
Change-Id: Ic4dbd14cab340a4569e0bc92fc1cab20381b8a99