Merge unused ext.cite.visualEditor.core ResourceModule away
This extra module is never used separately for anything. The only
reason for it to exist might be to make some kind of separation in
the code more visible. However:

1. We are told we shouldn't use RL modules for this, but keep their
number to an absolute minimum. Separating code can still be done via
directories, file names, and such.

2. Personally I don't find this separation very helpful. It appears
like the core module is (mostly) about "dm" and "ce" classes, and
the second module about "ui" classes. But this separation is not
clean (why is there a single "ui" class in core?) and not really
needed because we can tell the same from the class names alone.

The seperation exists since 2014, see da5eacbe. It was apparently
done for mobile. This is obsolete anyway by now because we are
moving mobile away from being something different. The two modules
load unconditionally since 2020, see I974654d.

Bug: T400800
Change-Id: I8dbbccfc4a0a71a537a442ee4c044c1a0ba336f4