Use ResourceLoader packageFiles instead of scripts
Note this is only one step of a multi-step process. Some of the
next steps are explained in comments in the code.

This patch intentionally leaves all the global, namespaced ve.…
class names untouched. This is done for two reasons: To make this
patch smaller and easier to review, and because some of these are
actually used in other codebases. These need to be updated first
to use `require` before we can remove the classes from ve.….

Bug: T400800
Change-Id: Id00cac82ad3bec8f89feb9e7f3792d057b1682c6