Split multi-class file MWReferenceDialogTool
This is required before we can turn these into packageFiles. One
class per file.

Note: The original file also contains a whole bunch of extra
registration stuff at the bottom. We might move this later. For the
moment it can stay, especially because it is not only related to
the class but all about that exact same "reference" tool.

Bug: T400800
Change-Id: I894cec782360955203cf4c84c14ab6eadcc7aef4