Save synthetic main body content on other refs
I extended the check on main nodes that might get the body content
to also take into account that the content could be "lost" on a
synthetic main ref that's not used anywhere.

Now any other ref can take that content to make sure it's not lost.

This unblocks all cases there's a "orphaned" synthetic main ref as
long as there's another refence with that key.

Bug: T391521
Change-Id: I02f590d5e7b102b63e60362f32f7952c6300e6b5