Remove a few not needed function closures from .js files
Looks like these closures are still needed in some cases. At least
for the two "init" files.

I expected QUnit tests to be automatically wrapped in a closure when
they are executed. But it appears like this is not the case. Their
global scope bleeds into the next test when I try to remove it.

Bug: T400800
Change-Id: I3bf0e8b51ded5448a99887ed106de5c4f61241cf