Parsoid: Don't remove the mainBody when matching with details
When we convert back to Wikitext we need to match a subref with
the main content if both should be merged into main+details.

There I removed code that removed the main mainBody element from
the DOM as soon as there's a match with details.  Another subref
that's parsed cannot find the main content then so we had weired
behavior.

Note, that the root problem could be having just one synthetic
main ref for both main+details here.  But even if we change that
in the future, the other mainBody element should have a different
id and therefore keeping both does not hurt.

Bug: T392816
Change-Id: Ic0c625c66fb11c2384f516fec4dacccf3523155f