fix: use MainConfig to access php config
The MediaWikiConfigReader is intended to access config stored in
CommunityConfiguration only. For a time, there was a deprecated
fallback. MediaWikiConfigRouter should be used if one wants to access
both config stored in CommunityConfiguration and PHP/extension.json.
When one wants to only access the config in PHP/extension.json, then
simply using MainConfig is the preferred way.

See change If0e5246578 in CommunityConfiguration that throws an
Exception if one tries to access PHP config through
MediaWikiConfigReader.

Bug: T402233
Bug: T387452
Change-Id: Ie6df26fd1fa71162d7170fb1ee8cd4864b6b6043