Displaying help dialog for add details button
We're adding a help dialog that will open when users click the
"Add details" button for the first time.  This is for providing
some basic onboarding.

Users can either close the dialog and trigger it again or dismiss
it to move on with adding details.  When dismissed the dialog will
not show up again.

The latter works using a hidden user preference or the local/cookie
storage of the browser.

Bug: T403693
Change-Id: If29c757eaa8f39df9b855b280156b5fb953000fe