VE: Better support LDRs that don't copy the group name
VE currently does not support list defined refs in group that don't
copy the group of the references tag.  I made two changes here to
fix that.

First I'm forwarding the group in Parsoid at a place where I think
it makes sense so that we can find it in VE.

Then I'm extracting the group on list defined refs in VE and set
the attribute accordingly.  That makes sure the internalItems are
set up correctly.

Finally we need to make sure that the group attribute from the node
does not get saved back to wikitext if it was not there before.
This seems easy because LDRs do not support changes to their group.

With the support of these LDRs we now allow that users edit them
including changing the group.  But I think that's a broken usecase
already.

Bug: T400596
Change-Id: I1cec066c959849d697a95ca61f4b5c05533a11a3