VE: Try to copy the node with reflistItemId when creating a synth ref
The synthetic refs need to have a reflistItemId that can be used as DOM
id to match the content from the referencesList to the main+details.

If there's a node with a reflistItemId use that one for the copy.

Note that this does not always work.  References that we add with
VE do not have a reflistItemId.

Bug: T397935
Change-Id: I1dba195f990148f08a2648c6ca55ac0211673372