VE: Generate a reflistItemId for synthetic refs if needed
When the user creats references in VE and then converts them to
sub-refs there's no reflistItemId that could be used to match
main and details in Parsoid later.

When this id is missing we need to generate it.  Parsoid does
that by calling ParsoidAnchorFormatter::getNoteIdentifier in PHP.

I'm not sure if it makes sense to try to rebuild that here.  At
least the id needs to be unique over all reflistItems.

Bug: T397935
Change-Id: I78a3312afa0854fee7331a3be9ddab5e49afa7f3