Remove meaningless href="#" from help panel examples
Modern browsers apparently don't need these. They just add confusion
and cause unwanted behavior when the fake links are clicked.

Unfortunately one of the message uses wikitext web links and cannot
be changed so easily. For the moment I suggest to change the anchor
so that the page doesn't jump all the way to the top but back to
the toolbar.

Bug: T399278
Change-Id: I035aedb873adc25cb78dd811a472b2b93f13a5c2