Re-define tab key in the reference reuse search dialog
The current behavior of the tab key in this dialog is that the
dialog looses focus. There is apparently nothing the user can do
then.

Instead, block the tab key from leaving the search results list and
highlight the first search result as a visual response for the user.
This makes it much more obvious that they need to use the cursor
keys to select one of the search results.

I considered suggesting this as an upstream change in OOUI. But it
feels a bit to heavy as it depends on the dialog and if there are
more elements after the search results in the dialog. Here we know
how the dialog works.

Bug: T403588
Change-Id: Icd1f7f5ca30f33a13b5a91f08162b006ed425f00