VE: Split reused sub-refs when editing only one of the reuses
At the moment we don't really support a real reuse off sub-refs.
We only want to implement that solution later with better UX
support.  Nevertheless we then also want to give the option to
edit either one reuse or split the sub-refs.

For split is the default.  So when editing a sub-ref with reuses
we create a new ref with new keys and replace the duplicate.

Bug: T404415
Change-Id: I1f81b1bddcad48302624b8a5f070b3cbd64e3320