VE: Code cleanup in sub-ref creation
I removed some deprecated code from an entrypoint that's not
relevant anymore.  Currently we only can add details when we've
selected a node.  The former entrypoint in the reuse tab's burger
menu got removed a while ago.  That old entrypoint could be used
to insert subrefs anywhere.  Now it's only possible when a ref
node is selected.

Bug: T404089
Change-Id: Ie5016c96117ac5d7696591e34e6cb21250744d3a