VE: Preserve contentsUsed flag when converting main refs
We want to make sure that main content does not move around in
Wikitext when saving back.

If a new sub-ref is replacing a main ref that had the content and
therefore the sub-ref should also become a main+details it needs
to get the contentsUsed flag to make sure that the sub-ref gets
the main content stored when converting back.

This patch is an in-between step to make sure the flag is set
correctly.

Bug: T404089
Change-Id: Ia515dc5111a7e82aaaa721bc022556281d30f393