VE: Use the contentsUsed flag to match sub-refs that store main content
To make sure that main content does not move around we need a flag
to remember which sub-ref was a main+details ref in Wikitext and
should therefore get the main content attached when converting back
to Wikitext.

The `contentsUsed` flag does exactly that for nodes that are pure
main refs.  For sub-refs that flag is meaningless at the moment
they should always get the body content attached because for these
nodes it represents the content of the details attribute.

So for sub-refs we can give the flag a new meaning to remember which
sub-ref also kept the main body in Wikitext.

For now I'm just refining the meaning of that flag.  We'll use it
later to figure out which node actually should be getting main
content.

Bug: T404089
Change-Id: If34ad196a63db323467deaabb626072ecb3a39c2