Replace ref content fail-safe with less aggressive one
Note these code paths can only be hit when the incoming HTML is
mangled or outdated and not compatible any more, e.g. when it comes
from a 3rd-party client that behaves different than VisualEditor.
Instead of dropping the nodes completely and losing the <ref> tag
completely we can at least save an empty <ref … /> with all the
original arguments.

This gives the user something to work with. They know the exact
position in the document where it happened, as well as the ref's
name. The spot will also be marked with an error message in almost
all cases (e.g. because the content got lost).

The current behavior is that the ref disappears with no trace, and
no error. This is usually not a good idea.

As discussed, the logging is unused at the moment and can easily be
added back when needed.

Bug: T404477
Change-Id: I82d2746e26e9cdeb18e6ed159c6e5e6a3c1dde97