VE Converter: Merge code that places the main content
We intrudced two mechanism that should take care of making sure
that main content is not lost when converting back to Wikitext.

One part took care of main refs and the other managed sub-refs.
Since this was not done in one place we needed to make sure that
one part takes precedence over the other.  Now that we track on
both kinds of nodes, which one should get the main content we can
also merged both paths to make sure that the flag is respected in
all cases.

To make that happen we need to consider a list of nodes that
includes main- and sub-references.

I'm also removing the former mechanism to link sub-refs to main
content that's obsolete now.

Bug: T404089
Change-Id: I44d84eb3ceac9689120531f1df606ae3adbc9379