Remove use of Parser::$extCite
Replace caching of Cite object on the Parser object with
a WeakMap in the factory and store it under the parser object.

Remove the ParserCloned hook, it is no longer needed.
On clone the old property was copy over to the new instance and this
hook has destroyed the reference to the old Cite object.
But the fresh parser does not has a Cite object in the cache

Bug: T343230
Change-Id: I5eca12f193d6d165d0179532a3918a6cbd675087