Remove problematic fallback behavior for Follow nodes
This reverts the code change from patch I4a983b0. Turns out we can
fix the incomplete test cases instead, as demonstrated by this
(minimal) patch.

The code is (in theory) problematic because it just picks the first
Follow node it can find, even if it's technically possible to have
multiple of them. Luckily this is not relevant in the real-world
scenarios the Follow feature is designed for. A real-world page will
never have multiple, conflicting Follow.

However, no need to stick to the problematic code when it's not
needed.

Bug: T404477
Change-Id: Ia11f56e3c0874bbff59a61a822f727510f8895d0