Fix inconsistent view/edit icon for subrefs
As discussed in I442297f.

1. In readonly mode we still see an edit pen, but the tooltip says
"view". This is a bit confusing. Let's use the icon that matches the
tooltip and the actual behavior.

2. We can remove the extra "mobile" class and use the currently
preferred mechanism to target a specific skin. I think that class was
added a while ago via I0b60306. Note we should generally not target
"mobile" any more but a skin.

Bug: T404690
Change-Id: I902d95c3b778d839f1fa3738c7f8c1c68062d261