Replace fragile "mobile" with Minerva skin targetting
Small code cleanups, as discussed in I5e69449. Notably: Use the
recommended CSS to target specific skins. Same as in I902d95c.

The fundamental problem here is: "Mobile" is not the same as
"Minerva". It's technically possible to have a mobile experience
in the Vector skin, or a desktop experiente in the Minerva skin.
However, the CSS in question here is not about the target, it's
about the skin and should say so.

Bug: T404690
Change-Id: Iccf2389a7a5086afbe5fa8b9f7c01a012dfa0a08