Only run VisualEditor QUnit tests when it is loaded
On CI, we would like to remove the recursive injection of extensions
dependencies. When Cite is a dependency, we are required to add
VisualEditor to have Cite QUnit tests to pass.

Make the tests optional based on whether VisualEditor is loaded.

Bug: T406323
Change-Id: I7b30ce9032e0598f62a06f61f507fdb47132e5fe
(cherry picked from commit 7816aaabc241873af9b4d4f827b74c2767a88190)