Fix misplaced (underlined) space in backlinks
This currently doesn't show up in visual diffs but makes a difference
when a user does have "always underline links" enabled. The underline
reaches to far and underlines some spaces in awkward places.

This is another minor discrepancy between the two parsers.

Note this is currently masked with some ::before CSS that replaces
the up-arrow that's actually in the HTML with an identical up-arrow
from CSS. You need to disable that CSS to see the difference. Or
look at the difference in the HTML.

Bug: T398402
Change-Id: If615a26a8faee12273f13aa925ecbf63980b2963