Start implementing sub-ref merging in legacy parser
This code is split from the later patch Iddadc58. The idea is to
prepare the relevant code to be able to render sub-refs as being
reused, without using that functionality by default.

What's critical for this patch is that none of the existing parser
tests changes, for none of the two parsers.

Bug: T385666
Bug: T406549
Change-Id: I960244789e6045f268a4c2bfaf2de7e48a51a95d