Remove unused, underspecified "public": true from extension.json
The "public" flag was an old idea that was never implemented. See
https://www.mediawiki.org/wiki/Manual:Extension_registration#public_/_private
I find it problematic because we don't even know what it means to
make these config flags "public", and if we really want to make them
all public or only a subset. That's a decision we can only make when
we know what it means. Until then we should rely on the default
(which is false).

It's probably obsolete anyway because of CommunityConfiguration.

Bug: T406549
Change-Id: I2b45de71d73cd4ffc2b920f8e6fad4fffcdd5084