Get the section preview option from parser options while parsing
Getting an option from ParserOptions is recorded and maybe used in the
parser cache key, avoid this side effect by removing
Cite::$isSectionPreview and not setting it in the constructor.
When a parse is not using this option it should not recorded.

Also remove the extra argument from public checkRefsNoReferences() to
ensure the correct option is used and it is not possible to use another
option, that is not recorded in the corresponding ParserOutput.

Change-Id: I4f63335d6087bb3a21d46945a78320323c1f5482