Much simpler/real-world examples in wikitext help panel
This came up while working on T399278. There are two problems with
these examples:

1. Two examples are almost identical. Both use the same link with
the same label. This makes the panel more confusing than it needs
to.

The wiki link might be useful to demonstrate that wikitext can be
used. But we don't need to repeat that a second time. Plain text is
enough.

2. The sub-ref example can be much closer to what this feature is
meant to represent in the real world. The example should be just
that, and example. The descripton is in the first column.

Bug: T407471
Change-Id: I98b4c5c5a0c2703f9f3792ba186e501c676fdb66