VE: Add MWReferenceModel QUnit test for internal item updates
There seems to be no distinct test for the `updateInternalItem`
method.  So I added one that tests changes on the group.

Group changes are relatively complex from what's happening in
the background to so this test is covering a lot of things.

Motivated by the failing tests after merging
I9e501ce1228edfefc01a1980b7670ad591f7ac69

Bug: T408535
Change-Id: Ic7117373f5ec89301d3b88dbbc67be191045afc0