Only enable VisualEditor integration when it is loaded
On CI, we would like to remove the recursive injection of extensions
dependencies. When Cite is a dependency, we are required to add
VisualEditor to fulfil MediaWiki ResourcesTest::testValidDependencies
which otherwise would fail with:

  Dependencies that do not exist
  Failed asserting that two arrays are equal.
  --- Expected
  +++ Actual
  @@ @@
   Array (
       'ext.cite.visualEditor' => Array (
  +        0 => 'ext.visualEditor.base'
  +        1 => 'ext.visualEditor.mediawiki'
  +        2 => 'ext.visualEditor.mwcore'
  +        3 => 'ext.visualEditor.mwtransclusion'
       )
   )

Move the `ext.cite.VisualEditor` registration from `extension.json` to
the PHP `ResourceLoaderRegisterModules` hook. Only load it when
VisualEditor is loaded.

Bug: T406323
Change-Id: I5b3bf13fbf5e0bae6ba34f939cd658eafd95df79