VE: Simplify ve.dm.Transaction test runner
I mainly did these things in this change to improve the readability
of this QUnit test:

A few annotations, better variable names, less complex indentation.

Removal of dead code in the "test runner" part.  The only tests
case in this file did neither test the removal of data nor the
data in the hash store.  So both parts of the test runner can be
removed.

In another step I might reduce the complexity of the test case
itself.  But I think, this is a good start already to better get
what's going on.

Motivated by the failing tests after merging
I9e501ce1228edfefc01a1980b7670ad591f7ac69

Bug: T408535
Change-Id: Ie242b53268449c4b1203d25d2cb955cd5e963648