VE: Reuse existing simple fixture in ve.dm.Transaction test
I reused the simple test data that I created for another test in
the Transacion test.  The differences between these two fixtures
are just minimal.

To avoid renumbering line numbers in the Transaction test I just
added the `Bar` text to the test data.  Because of this I also
needed to change the access to the node data in the other test.

Bug: T408535
Change-Id: Ic434b55f61a4e20e81e7cd2713d96ca48ff35f10