VE: Fix internal item data fixture in ve.dm.Transaction test
The `internalItems` property on the test data fixtures is added to
make sure, that the internalItems are populated correctly.

The values are used in ve.dm.example.createExampleDocumentFromData
as arguments to call ve.dm.InternalList.prototype.queueItemHtml

queueItemHtml does not allow any of it's values to be null so I
replaced it with the actuall listKey of the reference.

Also is the expected format of the group string to include the
slash at the end of `mwReferences` when it's the default group.

And I used the right format for the listKey that always includes
either `literal/` or `auto/`.

Bug: T408535
Change-Id: I5c76e1a201bbf6a301f75d1c67110bc7979b0362