VE: Change ve.dm.Transaction.test to cover input from copy and paste
I did a complex investigation why this test failed after a change
in the VE lib affecting the InternalList.  See T408535

To me it seems that the setup tested here is not covering any
actual use cases.  The method tested is quite complex and involves
a lot of code but the main use case seems to be copy and paste,
merging an external doc using the ve.ce.ClipboardHandler.

In that case remapInternalListKeys() is called and then the
expected outcome can be achieved.

Bug: T408535
Change-Id: I75cc6896aa034e0388864a8ff13b4cf64dd5e23f